@@ -15,20 +15,18 @@
     "bs58check": "2.1.2",
     "classnames": "2.5.1",
     "webextension-polyfill": "0.12.0",
-    "i18next": "19.8.4",
+    "i18next": "25.7.3",
     "libsodium-wrappers-sumo": "0.7.15",
     "mina-signer": "3.1.0",
     "@zondax/ledger-mina-js": "0.1.1",
     "zustand":"5.0.9",
-    "react": "17.0.1",
-    "react-circular-progressbar": "2.2.0",
-    "react-dom": "17.0.1",
-    "react-helmet": "6.1.0",
-    "react-i18next": "11.8.2",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
+    "react-i18next": "16.5.0",
     "react-idle-timer": "5.7.2",
-    "react-redux": "7.2.2",
-    "react-router-dom": "5.1.2",
-    "redux": "4.0.5",
+    "react-redux": "9.2.0",
+    "react-router-dom": "7.11.0",
+    "react-circular-progressbar": "2.2.0",
     "buffer": "6.0.3",
     "uuid": "11.1.0",
     "valid-url": "1.0.9",
