@@ -1,6 +1,6 @@
 import browser from 'webextension-polyfill';
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { NET_CONFIG_VERSION } from "../config";
 import { windowId } from "./background/DappService";
@@ -197,14 +197,15 @@ export const applicationEntry = {
   },
 
   render() {
-    ReactDOM.render(
+    const container = document.getElementById("root");
+    const root = createRoot(container);
+    root.render(
       <React.StrictMode>
         <Provider store={store}>
           <PopupMonitor />
           <App />
         </Provider>
-      </React.StrictMode>,
-      document.getElementById("root")
+      </React.StrictMode>
     );
   },
 };
