@@ -1,17 +1,17 @@
 import i18n from "i18next";
 import { useCallback } from "react";
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import CustomView from "../../component/CustomView";
 import styles from "./index.module.scss";
 
 
 const Security = ({ }) => {
-  const history = useHistory()
+  const navigate = useNavigate()
   const onToSeedPage = useCallback(() => {
-    history.push('reveal_seed_page')
+    navigate('/reveal_seed_page')
   }, [])
   const onResetPwd = useCallback(() => {
-    history.push('reset_password')
+    navigate('/reset_password')
   }, [])
   return <CustomView title={i18n.t('security')} contentClassName={styles.container}>
     <RowItem title={i18n.t('backupMnemonicPhrase')} onClickItem={onToSeedPage} />
